requirements.txt
transformers==4.36.2
torch
python-docx
python-pptx
PyPDF2
pdfplumber
streamlit